# PERF: DataFrame.values