```
from asv_bench.benchmarks.frame_methods import *
self = ToNumpy()
self.setup()

%timeit self.time_to_numpy_mixed_tall()
24.2 ms ± 324 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
11.8 ms ± 609 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```