@jreback 2021-08-30
cool maybe have a perf improvement note listing a bunch of issues would be good